issue #175: fix linting issues
