+++ internal/pkg/fs/fs.go
@@ -398,57 +398,6 @@ func (fs *virtualFileSystem) Lstat(name string) (os.FileInfo, error) {
 	return n, nil
 }
 
-func (fs *virtualFileSystem) mkdirCommon(name string, perm os.FileMode, all bool) error {
-	if (perm & os.ModeType) != 0 {
-		return errBadMode
-	}
-	n, nameRem, err := fs.find(name, false, true)
-	if err != nil && !os.IsNotExist(err) {
-		return err
-	}
-	if !all {
-		slashPos := strings.IndexByte(nameRem, '/')
-		if slashPos >= 0 {
-			return os.ErrNotExist
-		}
-		if nameRem == "" {
-			return os.ErrExist
-		}
-	}
-	if !n.mode.IsDir() {
-		return syscall.ENOTDIR
-	}
-	if all {
-		for nameRem != "" {
-			slashPos := strings.IndexByte(nameRem, '/')
-			nameComp := nameRem
-			if slashPos >= 0 {
-				nameComp = nameComp[:slashPos]
-			}
-			if nameComp != "" {
-				validateNameComp(nameComp)
-				childN := &node{
-					mode: perm | os.ModeDir,
-					name: nameComp,
-				}
-				n.dirAppend(childN)
-				n = childN
-			}
-			if slashPos < 0 {
-				nameRem = ""
-			} else {
-				nameRem = nameRem[slashPos+1:]
-			}
-		}
-		return nil
-	}
-	n.dirAppend(&node{
-		mode: perm | os.ModeDir,
-		name: nameRem,
-	})
-	return nil
-}
-
 func (fs *virtualFileSystem) Mkdir(name string, perm os.FileMode) error {
 	return fs.mkdirCommon(name, perm, false)
 }
